Updated branch

+++ backend/routes/api/members.js
@@ -7,6 +7,9 @@ require("dotenv").config();
 // Load input validation
 const validateRegisterInput = require("../../validation/register");
 const validateLoginInput = require("../../validation/login");
+const validateDemoteUser = require("../../validation/demote");
+const validateDeleteUser = require("../../validation/delete");
+const validatePromoteUser = require("../../validation/promote");
 
 // Load User model
 const Member = require("../../models/Member");
@@ -15,7 +18,6 @@ const Member = require("../../models/Member");
 // @desc Return all registered members
 // @access Public
 router.route("/").get(function (req, res) {
-  //res.set("Content-Type", "application/json");
   Member.find(function (err, members) {
     if (err) {
       res.json(err);
@@ -29,36 +31,32 @@ router.route("/").get(function (req, res) {
 // @desc Promotes a member to the next memberType
 // @access Admin
 router.put("/promote", (req, res) => {
-  if (req.body.memberType === "admin")
-    res.status(400).json("Unable to promote member above Admin");
-  else {
-    let newType = "";
-    if (req.body.memberType === "pending") newType = "member";
-    else if (req.body.memberType === "member") newType = "exec";
-    else if (req.body.memberType === "exec") newType = "admin";
-    else newType = "member";
+  const { errors, isValid, newType } = validatePromoteUser(req.body);
 
-    Member.findOneAndUpdate(
-      { email: req.body.email },
-      { memberType: newType },
-      (err) => {
-        if (err) {
-          res.status(400).send(err);
-        } else res.status(200).json("Member promoted to " + newType);
-      }
-    );
+  if(!isValid) {
+    return res.status(400).json(errors);
   }
+  
+  Member.findOneAndUpdate(
+    { email: req.body.email },
+    { memberType: newType },
+    (err) => {
+      if (err) {
+        res.status(400).send(err);
+      } else res.status(200).json("promoted to " + newType);
+    }
+  );
 });
 
 // @route PUT api/members/demote
 // @desc Demotes a member to the previous memberType
 // @access Admin
 router.route("/demote").put(function (req, res) {
-  let newType = "";
-  if (req.body.memberType === "pending") newType = "pending";
-  else if (req.body.memberType === "member") newType = "pending";
-  else if (req.body.memberType === "exec") newType = "member";
-  else newType = "admin";
+  const { errors, isValid, newType } = validateDemoteUser(req.body);
+  
+  if(!isValid){
+    return res.status(400).json(errors);
+  }
 
   Member.findOneAndUpdate(
     { email: req.body.email },
@@ -66,7 +64,7 @@ router.route("/demote").put(function (req, res) {
     (err) => {
       if (err) {
         res.status(400).send(err);
-      } else res.status(200).json("Member demoted to " + newType);
+      } else res.status(200).json("demoted to " + newType);
     }
   );
 });
@@ -75,6 +73,12 @@ router.route("/demote").put(function (req, res) {
 // @desc Deletes specified member
 // @access Public
 router.post("/", (req, res) => {
+  const { errors, isValid } = validateDeleteUser(req.body);
+
+  if(!isValid){
+    return res.status(400).json(errors);
+  }
+
   Member.findOneAndDelete({ email: req.body.email }, (err, member) => {
     if (err) {
       res.status(400).send(err);
